Merging with document and body tags.
Problems:
SlotImmutableException on copyNode method in merging. try/catch handling for now.

Problem to generalize. When Parsing version one and version two. For example, if we have a text tag on both version 0 and version 1. When parsing version 1 is not possible to recover the node that was created when parsing version 0.